Extract ConsoleColorizer from ColoredConsoleTarget.
